% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion-matrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{Confusion matrix}
\usage{
confusionMatrix(original_data, rule_data, target, inspection)
}
\arguments{
\item{original_data}{\code{dataframe} with data we are analizing}

\item{rule_data}{\code{dataframe} with subset according to subgroup conditions from \code{\link{ruleDataset}}}

\item{target}{character with target attribute value}

\item{inspection}{logical, if we apply confusion matrix after analysis for deper inspection of extracted subgroups, i.e.
can be aplied after evolutionary subgroups extraction \code{\link{subgroupDiscovery}}}
}
\value{
\code{matrix} with quality metrics
}
\description{
Confusion matrix in dimension 2 by 2
}
\details{
Compute confusion matrix dimensions 2x2. Matrix is presented like:
\tabular{ccc}{
---------------\tab Predicted: positive value \tab Predicted: negative value(s)\cr
Real: positive value\tab TP \tab FN\cr
Real: negative value(s)\tab FP \tab TN\cr
}
Where 1st value is target attribute value user specified for analysing. TP are true positives,
TN are true negative, FP are false positives and FN are false negatives. If target attribute has
more than one value positive value present specified target value and negative values all other values
to form matrix 2x2. If user does not specify target value, confusion matrix addapt for each value as stated
before (positive versus others). Calculated quality statistics for subgroup(s) are:
\itemize{
\item rule SIZE is number of instaces covered by rule conditions
\item true positives (TP) are number of positive instances covered by rule or specified target attribute value
\item COVERAGE as \eqn{rule_size / data_size}
\item SUPPORT as \eqn{tp / data_size}
\item positive predictive value (PPV) as \eqn{tp / rule_size}
\item SENSITIVITY or recall as \eqn{tp / (tp + fn )}
\item SPECIFICITY as \eqn{tn / (tn + fp)}
\item F-1 SCORE as \eqn{2*tp / (2 * tp + fp + fn)}
}

\code{inspection} argument serve for needs in post extraction, deper inspection of selected subgroup. Function
can be used for both analysis principles as soon we do analysis with \code{\link{subgroupDiscovery}} function,
because of unified subgroup presentation in form \code{attribute=value} separated by comma.
}
