% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDEFSR-analisis.R
\name{dataPartition}
\alias{dataPartition}
\title{Data partition}
\usage{
dataPartition(data_config, split, seed, class_vec)
}
\arguments{
\item{data_config}{\code{list} object with parameters for \code{\link[SDEFSR]{SDEFSR_DatasetFromDataFrame}}}

\item{split}{numeric, in range [0, 1]. Share for \code{training} and rest for \code{test} subset}
}
\value{
SDEFSR objects \code{list} with \code{training} set on 1st place and \code{test} on 2nd place
}
\description{
Split data into \code{training} and \code{test} subset. Class distribution is preserved.
}
\details{
Function split data and preserve target attribute value distribution.
After spliting on 2 parts, function create 2 \code{SDEFSR} object.
}
\seealso{
\code{\link[SDEFSR]{SDEFSR_DatasetFromDataFrame}}
}
