% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule-subdataset.R
\name{dynamicFilter}
\alias{dynamicFilter}
\title{Dynamic filter}
\usage{
dynamicFilter(dataset, columns, values, add_minors = FALSE,
  analysis = FALSE)
}
\arguments{
\item{dataset}{\code{dataframe} object, database on which analysis was perfomed}

\item{columns}{character \code{vector} with columns name}

\item{values}{character \code{vector} with values of selected columns}

\item{add_minors}{logical, default FALSE}

\item{analysis}{logical, default FALSE}
}
\value{
subset \code{dataframe}
}
\description{
Filter data by attribute (column) value
}
\details{
Filter database by variable value. After applied filter instances of minor target attribute value are added
in case of imbalace data or keep all target value occurences in subset data. If subset data is used for
further analysis columns on which filter have been applied are removed.
}
