% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDEFSR-analisis.R
\name{evolutionAnalysis}
\alias{evolutionAnalysis}
\title{Evolution analysis}
\usage{
evolutionAnalysis(data_config, method, ...)
}
\arguments{
\item{data_config}{list object with parameters for \code{\link[SDEFSR]{SDEFSR_DatasetFromDataFrame}}}

\item{method}{character. Name of algorithms: \code{\link[SDEFSR]{MESDIF}}, \code{\link[SDEFSR]{SDIGA}},
\code{\link[SDEFSR]{NMEEF_SD}}, \code{\link[SDEFSR]{FUGEPSD}}}

\item{...}{arguments that method accepts}
}
\value{
data frame with subgroups summary statistic for one target variable or list of data frames with subgroup
summary statistic for all target variables
}
\description{
Make subgroup discovery with evolutionary fuzzy logic algoritms implemented in SDEFSR package
}
\details{
Function that do subgroup analysis with SDEFSR package. User can use all 4 implemented evolutionary algorithms.
Arguments for selected method can be passed in as list or each individually. User has to specify correct name as
described in documentation of each method. Function does not return SDEFSR object but rather subgroups with
summary statistic. Addition to SDEFSR summary statistic for extracted subgroups are: size of subgroup and target count.
Function does not privide ability to calculate summary statistic from confusion matrix.
}
\seealso{
\code{\link[SDEFSR]{SDEFSR_DatasetFromDataFrame}}, \code{\link[SDEFSR]{MESDIF}}, \code{\link[SDEFSR]{SDIGA}},
\code{\link[SDEFSR]{NMEEF_SD}}, \code{\link[SDEFSR]{FUGEPSD}}
}
