% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result-summary.R
\name{evolutionSummaryTable}
\alias{evolutionSummaryTable}
\title{Evolutionary summary table}
\usage{
evolutionSummaryTable(original_data, SDEFSR_output)
}
\arguments{
\item{original_data}{data frame object with data on which analysis was perfomed}

\item{SDEFSR_output}{data frame object with results from SDEFSR package functions}
}
\value{
data frame with quality measures or list with data frame quality measure for each target attribute value
}
\description{
Get quality measures for SDEFSR package rules in data frame object.
}
\details{
Save calculated quality measures inside data frame object. Function used for rule extraction are:
\code{\link[SDEFSR]{FUGEPSD, MESDIF, NMEEF_SD, SDIGA}}. For now only MESDIF and NMEEF_SD are used.
}
