% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result-summary.R
\name{rsubgroupSummaryTable}
\alias{rsubgroupSummaryTable}
\title{Rsubgroup summary table}
\usage{
rsubgroupSummaryTable(original_data, rsubgroup, target)
}
\arguments{
\item{original_data}{\code{dataframe} object with database on which analysis was perfomed}

\item{rsubgroup}{\code{dataframe} object with results from \code{\link[rsubgroup]{DiscoverSubgroups}} function}

\item{target}{character with target attribute value}
}
\value{
data frame object
}
\description{
Calculate summary table for rsubgroup package subgroups.
}
\details{
Save calculated quality measures inside data frame object. Function can be used on output from
\code{\link[rsubgroup]{DiscoverSubgroups}}.
}
