% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule-subdataset.R
\name{ruleDataset}
\alias{ruleDataset}
\title{Subgroup subset}
\usage{
ruleDataset(subgroup, original_data)
}
\arguments{
\item{subgroup}{character \code{vector} of rule description from applied subgroup method output from
\code{\link{subgroupDiscovery}}.}

\item{original_data}{\code{dataframe} object with initial data for analysis}
}
\value{
\code{dataframe} with subset(s) of a rule(s)
}
\description{
Subset according to subgroup conditions
}
\details{
Function filter dataset for analysis according to conditions in extracted rules. Function can
extract subset for both approach: rsugroup package and evolutionary approach. Subgroups
conditions has to be outputed from \code{\link{subgroupDiscovery}} and then analyse each subgroup
in details if necessary. Function can be used with \code{\link[base]{lapply}}.
}
