% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sortDF}
\alias{sortDF}
\title{Sort dataframe}
\usage{
sortDF(df, by, decrease = TRUE)
}
\arguments{
\item{df}{\code{dataframe} to sort}

\item{by}{character, name of column}

\item{decrease}{logical, default TRUE}
}
\value{
Sorted \code{dataframe}
}
\description{
Sort \code{dataframe} by selected column in descreasing order
}
