% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sortL}
\alias{sortL}
\title{Sort list}
\usage{
sortL(li, by, decrease = TRUE)
}
\arguments{
\item{li}{\code{list} object with data frame as each element}

\item{by}{character, name of column}

\item{decrease}{logical, default TRUE}
}
\value{
Sorted \code{list} of \code{dataframes}
}
\description{
Sort \code{list} by selected column in descreasing order
}
