% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sortLL}
\alias{sortLL}
\title{Sort list of lists}
\usage{
sortLL(ll, by, decrease = TRUE)
}
\arguments{
\item{ll}{list of lists object with data frame as each element of innermost list}

\item{by}{character, name of column}

\item{decrease}{logical, default TRUE}
}
\value{
Sorted list of lists with data frames
}
\description{
Sort list by selected column in descreasing order
}
