% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result-summary.R
\name{subgroupSummary}
\alias{subgroupSummary}
\title{Subgroup summary tables}
\usage{
subgroupSummary(original_data, output, target = NULL, evolution = FALSE)
}
\arguments{
\item{original_data}{\code{dataframe} object with data on which analysis was perfomed}

\item{output}{output from algorithms for subgroup analysis in R}

\item{target}{character target attribute variable value, default NULL}

\item{evolution}{logical, default is FALSE.}
}
\value{
\code{dataframe} with subgroup and summary statistic
}
\description{
Calculate quality measures for extracted rules in subgroup discovery analysis
}
\details{
Summary statistic for each extracted rule from rsubgroup package and
evolutionary algorithms (SDEFSR) package. RSubgroup summary tables are calculated from confusion
matrix \code{\link{results_quality}}, while function add to summary static from SDEFSR package:
size of subgroup and target attribute value count. Argument output must be a output from:
\code{\link[rsubgroup]{DiscoverSubgroups}} or from \code{\link[SDEFSR]{MESDIF}}, \code{\link[SDEFSR]{SDIGA}},
\code{\link[SDEFSR]{NMEEF_SD}}, \code{\link[SDEFSR]{FUGEPSD}}.
}
