% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VREfmData-doc.R
\docType{data}
\name{VREfm}
\alias{VREfm}
\title{Vancomycin resistant Enteroccocus faecium (VREfm) database}
\format{A dataframe (201, 23)}
\usage{
VREfm
}
\description{
Patients with suceptibility testing for VREfm in University hospital from years 2014-1015
}
\keyword{datasets}
